!@ Tutorials Completed

+++ app/Http/Controllers/PostsController.php
@@ -3,10 +3,20 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use  Illuminate\Support\Facades\Storage;
 use App\Post;
 
 class PostsController extends Controller
 {
+    /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
     /**
      * Display a listing of the resource.
      *
@@ -41,11 +51,30 @@ class PostsController extends Controller
     {
         $this->validate($request, [
             'title' => 'required',
-            'body' => 'required'
+            'body' => 'required',
+            'cover_image' => 'image|nullable|max:1999'
         ]);
+
+        //Handle file Upload
+        if ($request->hasFile('cover_image')) {
+            //Get File Name With Extension
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            //Get Filename
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+            //Get Extension
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            //Get File Name With Extension
+            $fileNameToStore = $filename.'_'.time().'_'.$extension;
+            //Upload Image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+        } else {
+            $fileNameToStore = 'noimage.jpg';
+        }
+
         $post = new Post;
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        $post->cover_image = $fileNameToStore;
         $post->user_id = auth()->user()->id;
         $post->save();
 
@@ -73,6 +102,11 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+
+        // check for correct user
+        if (auth()->user()->id !== $post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthorized Access');    
+        }
         return view('posts.edit')->with('post',$post);
     }
 
@@ -89,9 +123,27 @@ class PostsController extends Controller
             'title' => 'required',
             'body' => 'required'
         ]);
+
+        //Handle file Upload
+        if ($request->hasFile('cover_image')) {
+            //Get File Name With Extension
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            //Get Filename
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+            //Get Extension
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            //Get File Name With Extension
+            $fileNameToStore = $filename.'_'.time().'_'.$extension;
+            //Upload Image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+        }
+
         $post = Post::find($id);
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        if ($request->hasFile('cover_image')) {
+            $post->cover_image = $fileNameToStore;
+        }
         $post->save();
 
         return redirect('/posts')->with('success', 'Post Updated');
@@ -106,6 +158,16 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+
+        // check for correct user
+        if (auth()->user()->id !== $post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthorized Access');    
+        }
+
+        if ($post->cover_image != 'noimage.jpg') {
+            //Delete
+            Storage::delete('public/cover_images/'.$post->cover_image);
+        }
         $post->delete();
         return redirect('/posts')->with('success', 'Post Deleted');
     }
